requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
openai
